Only try to use the outlets after viewDidLoad
When awakeFromNib is called, the view controller is done unpacking,
but the view hasn't been loaded yet, and so neither are the table
view cell outlets connected.

Fixes bug where if you set the map mode to hybrid and quit the app,
the settings panel will show it as "standard" instead.